refactor: move `query-to-params` utility to `utils` and export it directly from the SDK index.
