+++ sdk/src/index.ts
@@ -8,3 +8,4 @@ export * from './rest/index.js';
 export * from './schema/index.js';
 export type * from './types/index.js';
 export * from './utils/is-directus-error.js';
+export * from './utils/query-to-params.js';
