+++ sdk/src/rest/helpers/with-search.ts
@@ -1,4 +1,4 @@
-import { formatFields } from '../index.js';
+import { formatFields } from '../../utils/query-to-params.js';
 import type { RestCommand } from '../types.js';
 
 export function withSearch<Schema, Output>(getOptions: RestCommand<Output, Schema>): RestCommand<Output, Schema> {
