+++ sdk/src/rest/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './query-to-params.js';
 export * from './throw-if-empty.js';
 export * from './throw-core-collection.js';
 export * from './get-auth-endpoint.js';
